requirements.txt
django
djangorestframework
environs
mysqlclient
google-generativeai